Moving Disk buffering config to its feature dir
