+++ instrumentation/src/main/java/io/opentelemetry/android/config/OtelRumConfig.java
@@ -6,6 +6,7 @@
 package io.opentelemetry.android.config;
 
 import io.opentelemetry.android.ScreenAttributesSpanProcessor;
+import io.opentelemetry.android.features.diskbuffering.DiskBufferingConfiguration;
 import io.opentelemetry.android.instrumentation.network.CurrentNetworkProvider;
 import io.opentelemetry.api.common.Attributes;
 import java.util.function.Supplier;
